# Optimize Mysql56GTIDSet.String(), add benchmark