## Description

This PR optimizes the `.String()` method of `mysql.Mysql56GTIDSet` by appending the byte-representation of the GTID-SID to the `bytes.Buffer` vs converting and writing it to the buffer as a `string`. Also a benchmark test for the `.String()` method was added

I also attempted to optimize further by writing the `start`/`end` intervals _(`int64`s)_ to the `bytes.Buffer` using `binary.Write()`, but confusingly, this performed worse than this change and I can't explain why 😕

In benchmarks this PR improves the efficiency of `.String()` by about 27%: `200.9 ns/op` vs `275.5 ns/op`

#### Before benchmark
```bash
$ go test -v -run BenchmarkMysql56GTIDSetString -bench=BenchmarkMysql56GTIDSetString ./go/mysql/...
goos: darwin
goarch: amd64
pkg: [vitess.io/vitess/go/mysql](http://vitess.io/vitess/go/mysql)
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkMysql56GTIDSetString
BenchmarkMysql56GTIDSetString-16    	 4378540	       275.5 ns/op
PASS
```

#### After benchmark
```bash
$ go test -v -run BenchmarkMysql56GTIDSetString -bench=BenchmarkMysql56GTIDSetString ./go/mysql/...
goos: darwin
goarch: amd64
pkg: [vitess.io/vitess/go/mysql](http://vitess.io/vitess/go/mysql)
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkMysql56GTIDSetString
BenchmarkMysql56GTIDSetString-16    	 6013492	       200.9 ns/op
PASS
```

If possible, please backport this to v13+ 🙇 

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported
-   [x] Tests were added or are not required
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->